fix: Addressed to-dos 1-3 of previous commit
Big change was using PKCE to retrieve the access code. WHat is in the url is now an authorization code. Also, this source helped to debug an issue where the request to get an access token required a client secret (although this application doesn't store that): https://stackoverflow.com/questions/76528208/google-oauth-2-0-authorization-code-with-pkce-requires-a-client-secret.